requirements.txt
Flask
requests
pandas
plotly
python-dotenv
